+++ src/common/Makefile.include
@@ -2,6 +2,7 @@
 #
 # Copyright (c) 2015-2019 Intel, Inc.  All rights reserved.
 # Copyright (c) 2016      Cisco Systems, Inc.  All rights reserved.
+# Copyright (c) 2023      Nanook Consulting.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -17,7 +18,9 @@ sources += \
         common/pmix_data.c \
         common/pmix_security.c \
         common/pmix_iof.c \
-        common/pmix_attributes.c
+        common/pmix_attributes.c \
+        common/pmix_alloc.c \
+        common/pmix_session.c
 
 headers += \
         common/pmix_iof.h \
