+++ src/common/pmix_query.c
@@ -680,353 +680,6 @@ PMIX_EXPORT pmix_status_t PMIx_Query_info_nb(pmix_query_t queries[], size_t nque
     return PMIX_SUCCESS;
 }
 
-static void acb(pmix_status_t status, pmix_info_t *info, size_t ninfo, void *cbdata,
-                pmix_release_cbfunc_t release_fn, void *release_cbdata)
-{
-    pmix_cb_t *cb = (pmix_cb_t *) cbdata;
-    size_t n;
-
-    cb->status = status;
-    if (NULL != info) {
-        PMIX_INFO_CREATE(cb->info, ninfo);
-        if (NULL == cb->info) {
-            cb->status = PMIX_ERR_NOMEM;
-            goto done;
-        }
-        cb->ninfo = ninfo;
-        for (n = 0; n < ninfo; n++) {
-            PMIX_INFO_XFER(&cb->info[n], &info[n]);
-        }
-    }
-
-done:
-    if (NULL != release_fn) {
-        release_fn(release_cbdata);
-    }
-    PMIX_WAKEUP_THREAD(&cb->lock);
-}
-
-PMIX_EXPORT pmix_status_t PMIx_Allocation_request(pmix_alloc_directive_t directive,
-                                                  pmix_info_t *info, size_t ninfo,
-                                                  pmix_info_t **results, size_t *nresults)
-{
-    pmix_cb_t cb;
-    pmix_status_t rc;
-
-    PMIX_ACQUIRE_THREAD(&pmix_global_lock);
-
-    if (pmix_globals.init_cntr <= 0) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_INIT;
-    }
-    PMIX_RELEASE_THREAD(&pmix_global_lock);
-
-    pmix_output_verbose(2, pmix_globals.debug_output, "%s pmix:allocate",
-                        PMIX_NAME_PRINT(&pmix_globals.myid));
-
-    /* set the default response */
-    *results = NULL;
-    *nresults = 0;
-
-    /* create a callback object as we need to pass it to the
-     * recv routine so we know which callback to use when
-     * the return message is recvd */
-    PMIX_CONSTRUCT(&cb, pmix_cb_t);
-    if (PMIX_SUCCESS != (rc = PMIx_Allocation_request_nb(directive, info, ninfo, acb, &cb))) {
-        PMIX_DESTRUCT(&cb);
-        return rc;
-    }
-
-    /* wait for the operation to complete */
-    PMIX_WAIT_THREAD(&cb.lock);
-    rc = cb.status;
-    if (NULL != cb.info) {
-        *results = cb.info;
-        *nresults = cb.ninfo;
-        /* protect the data */
-        cb.info = NULL;
-        cb.ninfo = 0;
-    }
-    PMIX_DESTRUCT(&cb);
-
-    pmix_output_verbose(2, pmix_globals.debug_output, "pmix:allocate completed");
-
-    return rc;
-}
-
-PMIX_EXPORT pmix_status_t PMIx_Allocation_request_nb(pmix_alloc_directive_t directive,
-                                                     pmix_info_t *info, size_t ninfo,
-                                                     pmix_info_cbfunc_t cbfunc, void *cbdata)
-{
-    pmix_buffer_t *msg;
-    pmix_cmd_t cmd = PMIX_ALLOC_CMD;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cb;
-
-    pmix_output_verbose(2, pmix_globals.debug_output, "pmix: allocate called");
-
-    PMIX_ACQUIRE_THREAD(&pmix_global_lock);
-
-    if (pmix_globals.init_cntr <= 0) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_INIT;
-    }
-
-    /* if we are the server, then we just issue the request and
-     * return the response */
-    if (PMIX_PEER_IS_SERVER(pmix_globals.mypeer) && !PMIX_PEER_IS_LAUNCHER(pmix_globals.mypeer)) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        if (NULL == pmix_host_server.allocate) {
-            /* nothing we can do */
-            return PMIX_ERR_NOT_SUPPORTED;
-        }
-        pmix_output_verbose(2, pmix_globals.debug_output, "pmix:allocate handed to RM");
-        rc = pmix_host_server.allocate(&pmix_globals.myid, directive, info, ninfo, cbfunc, cbdata);
-        return rc;
-    }
-
-    /* if we are a client, then relay this request to the server */
-
-    /* if we aren't connected, don't attempt to send */
-    if (!pmix_globals.connected) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_UNREACH;
-    }
-    PMIX_RELEASE_THREAD(&pmix_global_lock);
-
-    msg = PMIX_NEW(pmix_buffer_t);
-    /* pack the cmd */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &cmd, 1, PMIX_COMMAND);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-
-    /* pack the directive */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &directive, 1, PMIX_ALLOC_DIRECTIVE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-
-    /* pack the info */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &ninfo, 1, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-    if (0 < ninfo) {
-        PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, info, ninfo, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            PMIX_RELEASE(msg);
-            return rc;
-        }
-    }
-
-    /* create a callback object as we need to pass it to the
-     * recv routine so we know which callback to use when
-     * the return message is recvd */
-    cb = PMIX_NEW(pmix_query_caddy_t);
-    cb->cbfunc = cbfunc;
-    cb->cbdata = cbdata;
-
-    /* push the message into our event base to send to the server */
-    PMIX_PTL_SEND_RECV(rc, pmix_client_globals.myserver, msg, query_cbfunc, (void *) cb);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_RELEASE(msg);
-        PMIX_RELEASE(cb);
-    }
-
-    return rc;
-}
-
-
-static void blkcbfunc(struct pmix_peer_t *peer, pmix_ptl_hdr_t *hdr,
-                      pmix_buffer_t *buf, void *cbdata)
-{
-    pmix_shift_caddy_t *cd = (pmix_shift_caddy_t *) cbdata;
-    pmix_status_t rc, status;
-    int cnt;
-    PMIX_HIDE_UNUSED_PARAMS(hdr);
-
-    pmix_output_verbose(2, pmix_globals.debug_output,
-                        "pmix:query cback from server");
-
-    /* a zero-byte buffer indicates that this recv is being
-     * completed due to a lost connection */
-    if (PMIX_BUFFER_IS_EMPTY(buf)) {
-        return;
-    }
-
-    /* unpack the returned status */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &status, &cnt, PMIX_STATUS);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        status = rc;
-    }
-
-    /* if we have a cbfunc, execute it */
-    if (NULL != cd->cbfunc.opcbfn) {
-        cd->cbfunc.opcbfn(status, cd->cbdata);
-    }
-
-    /* cleanup */
-    PMIX_RELEASE(cd);
-}
-
-static void opcb(pmix_status_t status, void *cbdata)
-{
-    pmix_cb_t *cb = (pmix_cb_t*)cbdata;
-
-    cb->status = status;
-    PMIX_RELEASE_THREAD(&cb->lock);
-}
-
-PMIX_EXPORT pmix_status_t PMIx_Resource_block(pmix_resource_block_directive_t directive,
-                                              char *block, pmix_info_t *info, size_t ninfo)
-{
-    pmix_cb_t cb;
-    pmix_status_t rc;
-
-    PMIX_ACQUIRE_THREAD(&pmix_global_lock);
-
-    if (pmix_globals.init_cntr <= 0) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_INIT;
-    }
-    PMIX_RELEASE_THREAD(&pmix_global_lock);
-
-    pmix_output_verbose(2, pmix_globals.debug_output, 
-                        "%s pmix:resource block op",
-                        PMIX_NAME_PRINT(&pmix_globals.myid));
-
-    /* create a callback object as we need to pass it to the
-     * recv routine so we know which callback to use when
-     * the return message is recvd */
-    PMIX_CONSTRUCT(&cb, pmix_cb_t);
-    rc = PMIx_Resource_block_nb(directive, block, info, ninfo, opcb, &cb);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_DESTRUCT(&cb);
-        return rc;
-    }
-
-    /* wait for the operation to complete */
-    PMIX_WAIT_THREAD(&cb.lock);
-    rc = cb.status;
-    PMIX_DESTRUCT(&cb);
-
-    pmix_output_verbose(2, pmix_globals.debug_output, 
-                        "pmix:resource block operation completed");
-
-    return rc;
-}
-
-
-PMIX_EXPORT pmix_status_t PMIx_Resource_block_nb(pmix_resource_block_directive_t directive,
-                                                 char *block, pmix_info_t *info, size_t ninfo,
-                                                 pmix_op_cbfunc_t cbfunc, void *cbdata)
-{
-    pmix_buffer_t *msg;
-    pmix_cmd_t cmd = PMIX_RESBLK_CMD;
-    pmix_status_t rc;
-    pmix_shift_caddy_t *cb;
-
-    pmix_output_verbose(2, pmix_globals.debug_output, "pmix: allocate called");
-
-    PMIX_ACQUIRE_THREAD(&pmix_global_lock);
-
-    if (pmix_globals.init_cntr <= 0) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_INIT;
-    }
-
-    /* if we are the server, then we just issue the request and
-     * return the response */
-    if (PMIX_PEER_IS_SERVER(pmix_globals.mypeer) && !PMIX_PEER_IS_LAUNCHER(pmix_globals.mypeer)) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        if (NULL == pmix_host_server.resource_block) {
-            /* nothing we can do */
-            return PMIX_ERR_NOT_SUPPORTED;
-        }
-        pmix_output_verbose(2, pmix_globals.debug_output, "pmix:allocate handed to RM");
-        rc = pmix_host_server.resource_block(&pmix_globals.myid, directive, block,
-                                             info, ninfo, cbfunc, cbdata);
-        return rc;
-    }
-
-    /* if we are a client, then relay this request to the server */
-
-    /* if we aren't connected, don't attempt to send */
-    if (!pmix_globals.connected) {
-        PMIX_RELEASE_THREAD(&pmix_global_lock);
-        return PMIX_ERR_UNREACH;
-    }
-    PMIX_RELEASE_THREAD(&pmix_global_lock);
-
-    msg = PMIX_NEW(pmix_buffer_t);
-    /* pack the cmd */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &cmd, 1, PMIX_COMMAND);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-
-    /* pack the directive */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &directive, 1, PMIX_ALLOC_DIRECTIVE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-
-    /* pack the block name */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &block, 1, PMIX_STRING);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-
-    /* pack the info */
-    PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, &ninfo, 1, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(msg);
-        return rc;
-    }
-    if (0 < ninfo) {
-        PMIX_BFROPS_PACK(rc, pmix_client_globals.myserver, msg, info, ninfo, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            PMIX_RELEASE(msg);
-            return rc;
-        }
-    }
-
-    /* create a callback object as we need to pass it to the
-     * recv routine so we know which callback to use when
-     * the return message is recvd */
-    cb = PMIX_NEW(pmix_shift_caddy_t);
-    cb->cbfunc.opcbfn = cbfunc;
-    cb->cbdata = cbdata;
-
-    /* push the message into our event base to send to the server */
-    PMIX_PTL_SEND_RECV(rc, pmix_client_globals.myserver, msg, blkcbfunc, (void *) cb);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_RELEASE(msg);
-        PMIX_RELEASE(cb);
-    }
-
-    return rc;
-}
-
-
 /*
  * Determine if the keysis meant to be locally resolved
  */
